mutualization of source for os_info
