+++ nym-vpn-app/src-tauri/Cargo.toml
@@ -45,12 +45,11 @@ tower = "0.5.2"
 rust_iso3166 = "0.1"
 dirs = "6.0"
 semver = "1.0"
-sysinfo = "0.36.0"
 url = "2.5.4"
 sentry = { version = "0.42.0", features = ["tracing", "logs"] }
-sha2 = "0.10.9"
 
 # nym deps
+nym-platform-metadata = { path = "../../nym-vpn-core/crates/nym-platform-metadata" }
 nym-vpn-proto = { path = "../../nym-vpn-core/crates/nym-vpn-proto" }
 nym-ipc = { path = "../../nym-vpn-core/crates/nym-ipc" }
