+++ nym-vpn-app/src-tauri/src/sys.rs
@@ -1,9 +1,8 @@
+use nym_platform_metadata::SysInfo;
 use serde::Serialize;
-use sha2::{Digest, Sha256};
 use std::env;
 #[cfg(any(target_os = "linux", target_os = "openbsd"))]
 use std::process::Command;
-use sysinfo::System;
 #[cfg(any(target_os = "linux", target_os = "openbsd"))]
 use tracing::{error, info, warn};
 use ts_rs::TS;
@@ -61,18 +60,22 @@ pub struct OsInfo {
     pub display_server: DisplayServer,
     #[cfg(any(target_os = "linux", target_os = "openbsd"))]
     pub gpu: GpuType,
+    pub hasher: String,
 }
 
 impl OsInfo {
     pub fn new() -> Self {
+        let system = SysInfo::new();
+        let hasher = system.hash_identifier();
         Self {
-            version: System::long_os_version().unwrap_or_else(|| env::consts::OS.into()),
-            kernel: System::kernel_version(),
-            arch: env::consts::ARCH.to_string(),
+            version: system.os_version,
+            kernel: Some(system.kernel_version),
+            arch: system.arch,
             #[cfg(any(target_os = "linux", target_os = "openbsd"))]
             display_server: get_display_server(),
             #[cfg(any(target_os = "linux", target_os = "openbsd"))]
             gpu: gpu_info(),
+            hasher,
         }
     }
 
@@ -86,19 +89,6 @@ impl OsInfo {
                 env::set_var("WEBKIT_DISABLE_DMABUF_RENDERER", "1");
             }
         }
-    }
-
-    pub fn hash_identifier(&self) -> String {
-        let parts = [
-            self.version.clone(),
-            self.kernel.clone().unwrap_or_default(),
-            self.arch.clone(),
-            sysinfo::System::host_name().unwrap_or_else(|| "unknown".to_string()),
-        ];
-
-        let os_name = parts.join(" ");
-        let hash = Sha256::digest(os_name.as_bytes());
-        format!("{hash:x}")
     }
 }
