+++ nym-vpn-core/crates/nym-vpn-lib/src/platform/mod.rs
@@ -63,6 +63,7 @@ use std::{env, path::PathBuf, sync::Arc};
 use account::AccountControllerHandle;
 use lazy_static::lazy_static;
 use nym_gateway_directory::CachingGatewayClient;
+use nym_platform_metadata::SysInfo;
 use nym_vpn_api_client::types::ScoreThresholds;
 use sentry::ClientInitGuard;
 use tokio::{runtime::Runtime, sync::Mutex};
@@ -139,7 +140,7 @@ pub fn configureLib(config: NymVpnLibConfig) -> Result<(), VpnError> {
 
 async fn configure_lib(config: NymVpnLibConfig) -> Result<(), VpnError> {
     let network = environment::current_environment_details().await?;
-    let os = crate::SysInfo::new();
+    let os = SysInfo::new();
     os.raw_display(true);
     if config.sentry_monitoring {
         let mut guard = SENTRY_CLIENT.lock().await;
@@ -168,7 +169,7 @@ async fn init_logger(path: Option<PathBuf>, debug_level: Option<String>, sentry_
     let default_log_level = env::var("RUST_LOG").unwrap_or("info".to_string());
     let log_level = debug_level.unwrap_or(default_log_level);
     tracing::info!("Setting log level: {log_level}, path?: {path:?}");
-    let os = crate::SysInfo::new();
+    let os = SysInfo::new();
     os.display(true);
     if sentry_monitoring {
         let mut guard = SENTRY_CLIENT.lock().await;
