+++ nym-vpn-core/crates/nym-vpnd/Cargo.toml
@@ -45,7 +45,7 @@ sentry = { workspace = true, features = [
     "tracing",
     "logs",
 ] }
-sha2.workspace = true
+# sha2.workspace = true
 
 # Local crates
 nym-ipc.workspace = true
@@ -56,6 +56,7 @@ nym-vpn-api-client = { workspace = true }
 nym-vpn-lib = { workspace = true }
 nym-vpn-lib-types = { workspace = true }
 nym-vpn-network-config = { workspace = true }
+nym-platform-metadata = { workspace = true }
 nym-vpn-proto = { workspace = true, features = ["conversions"] }
 nym-vpn-store = { workspace = true }
 nym-vpnd-types = { workspace = true }
