+++ nym-vpn-core/crates/nym-vpnd/src/sentry.rs
@@ -2,7 +2,6 @@
 // SPDX-License-Identifier: GPL-3.0-only
 
 use sentry::{ClientInitGuard, Level};
-use sha2::{Digest, Sha256};
 use std::{borrow::Cow, sync::Arc, time::Duration};
 
 use crate::{config::GlobalConfigFile, environment};
@@ -26,7 +25,7 @@ pub fn init_sentry() -> Option<ClientInitGuard> {
         return None;
     };
 
-    let os_info = nym_vpn_lib::SysInfo::new();
+    let os_info = nym_platform_metadata::SysInfo::new();
 
     println!("Sentry monitoring enabled");
     let guard = sentry::init((
@@ -57,23 +56,11 @@ pub fn init_sentry() -> Option<ClientInitGuard> {
         scope.set_tag("os_version", &os_info.os_version);
         scope.set_tag("extra_metadata", os_info.extra.join(", "));
         scope.set_user(Some(sentry::User {
-            id: Some(anonymize_identifier(&os_info)), // anonymized user identifier
+            id: Some(os_info.hash_identifier()), // anonymized user identifier
             ip_address: None,
             ..Default::default()
         }));
     });
 
     Some(guard)
-}
-
-fn anonymize_identifier(os_info: &nym_vpn_lib::SysInfo) -> String {
-    let identifier = format!(
-        "{} {} {} {}",
-        os_info.os_version,
-        os_info.arch,
-        os_info.extra.join(" "),
-        sysinfo::System::host_name().unwrap_or_else(|| "unknown".to_string())
-    );
-    let hash = Sha256::digest(identifier.as_bytes());
-    format!("{hash:x}")
 }
